requirements.txt
Flask
flask-cors
Werkzeug
python-dotenv
pymongo
dnspython
requests
scikit-learn
joblib
numpy
pandas
scipy
tqdm
jinja2
markupsafe
click
itsdangerous
gunicorn